// types/DeliveryTypes.ts
export interface DeliveryTypes {
  id: string;
  name: string;
  email: string;
  phone: string;
  country: string;
  address: string;
  state: string;
  tracking_ref: string;
  delivery_time: string;
  status: string;
  delivery_type: string;
  certificate_id: string;
  user: string;
  created_at: string | Date;
  updated_at: string | Date;
}